fix README
