+++ .gitignore
@@ -6,6 +6,12 @@ __pycache__/
 # C extensions
 *.so
 
+# IDE
+.idea
+
+# Logs
+/logs
+
 # Distribution / packaging
 .Python
 build/
