+++ README.md
@@ -1,19 +1,16 @@
 Heart Disease UCI
 ==============================
 
-Example of ml project
-
-Installation: 
+Подготовка окружения: 
 ~~~
 python -m venv .venv
-source .venv/bin/activateW
+source .venv/bin/activate
 pip install -r requirements.txt
 ~~~
 Usage:
 ~~~
-python ml_example/train_pipeline.py configs/train_config.yaml
+python heart_disease/train_pipeline.py -d data/raw/heart.csv -c configs/train_config.yaml
 ~~~
-
 Test:
 ~~~
 pytest tests/
@@ -60,6 +57,27 @@ Project Organization
     └── tox.ini            <- tox file with settings for running tox; see tox.readthedocs.io
 
 
---------
+Критерии оценивания
+------------
 
-<p><small>Project based on the <a target="_blank" href="https://drivendata.github.io/cookiecutter-data-science/">cookiecutter data science project template</a>. #cookiecutterdatascience</small></p>
+-[x] Назовите ветку homework1 (1 балл) 
+-[x] положите код в папку src
+-[ ] В описании к пулл реквесту описаны основные "архитектурные" и тактические решения, которые сделаны в вашей работе. В общем, описание что именно вы сделали и для чего, чтобы вашим ревьюерам было легче понять ваш код. (2 балла)
+-[x] Выполнение EDA, закоммитьте ноутбук в папку с ноутбуками (2 баллов)
+    - Вы так же можете построить в ноутбуке прототип(если это вписывается в ваш стиль работы)
+    - Можете использовать не ноутбук, а скрипт, который сгенерит отчет, закоммитьте и скрипт и отчет (за это + 1 балл)
+-[x]  Проект имеет модульную структуру(не все в одном файле =) ) (2 баллов)
+-[x] Использованы логгеры (2 балла)
+-[x] Написаны тесты на отдельные модули и на прогон всего пайплайна(3 баллов)
+-[ ] Для тестов генерируются синтетические данные, приближенные к реальным (3 баллов)
+    - Можно посмотреть на библиотеки https://faker.readthedocs.io/en/, https://feature-forge.readthedocs.io/en/latest/
+    - Можно просто руками посоздавать данных, собственноручно написанными функциями
+    - Как альтернатива, можно закоммитить файл с подмножеством трейна(это не оценивается) 
+-[ ] Обучение модели конфигурируется с помощью конфигов в json или yaml, закоммитьте как минимум 2 корректные конфигурации, с помощью которых можно обучить модель (разные модели, стратегии split, preprocessing) (3 балла)
+-[x] Используются датаклассы для сущностей из конфига, а не голые dict (3 балла) 
+-[ ] Используйте кастомный трансформер(написанный своими руками) и протестируйте его(3 балла)
+-[ ] Обучите модель, запишите в readme как это предлагается (3 балла)
+-[ ] Напишите функцию predict, которая примет на вход артефакт/ы от обучения, тестовую выборку(без меток) и запишет предикт, напишите в readme как это сделать (3 балла)  
+-[ ] Используется hydra  (https://hydra.cc/docs/intro/) (3 балла - доп баллы)
+-[ ] Настроен CI(прогон тестов, линтера) на основе github actions  (3 балла - доп баллы (будем проходить дальше в курсе, но если есть желание поразбираться - welcome)
+-[ ] Проведите самооценку, опишите, в какое колво баллов по вашему мнению стоит оценить вашу работу и почему (1 балл доп баллы)
