+++ configs/logging.conf.yml
@@ -1,0 +1,41 @@
+version: 1
+consol_log:
+  level: DEBUG
+formatters:
+  simple:
+    class: logging.Formatter
+    format: "%(asctime)s %(levelname)s %(message)s"
+    datefmt: "%Y-%m-%d %H:%M:%S"
+handlers:
+  debug_handler:
+    class: logging.FileHandler
+    filename: logs/train-pipeline.log
+    level: DEBUG
+    formatter: simple
+  warn_handler:
+    class: logging.FileHandler
+    filename: logs/train-pipeline.warn
+    level: WARNING
+    formatter: simple
+loggers:
+  train_pipeline:
+    level: DEBUG
+    handlers: [debug_handler, warn_handler]
+    propagate: no
+root:
+  level: DEBUG
+  handlers: [debug_handler]
+
+capture: |2
+                              __  __          _____  ______
+                             |  \/  |   /\   |  __ \|  ____|
+                             | \  / |  /  \  | |  | | |__
+                             | |\/| | / /\ \ | |  | |  __|
+                             | |  | |/ ____ \| |__| | |____
+    _______ _____            |_|__|_/_/_   \_\_____/|______|_____  ______ _      _____ _   _ ______
+   |__   __|  __ \     /\   |_   _| \ | |      |  __ \_   _|  __ \|  ____| |    |_   _| \ | |  ____|
+      | |  | |__) |   /  \    | | |  \| |______| |__) || | | |__) | |__  | |      | | |  \| | |__
+      | |  |  _  /   / /\ \   | | | . ` |______|  ___/ | | |  ___/|  __| | |      | | | . ` |  __|
+      | |  | | \ \  / ____ \ _| |_| |\  |      | |    _| |_| |    | |____| |____ _| |_| |\  | |____
+      |_|  |_|  \_\/_/    \_\_____|_| \_|      |_|   |_____|_|    |______|______|_____|_| \_|______|
+
