+++ configs/train_config.yaml
@@ -2,38 +2,26 @@ input_data_path: "data/raw/train.csv"
 output_model_path: "models/model.pkl"
 metric_path: "models/metrics.json"
 splitting_params:
-  val_size: 0.1
-  random_state: 3
+  val_size: 0.25
+  random_state: 42
 train_params:
-  model_type: "RandomForestRegressor"
+  model_type: "LogisticRegression"
 feature_params:
   categorical_features:
-    - "MSZoning"
-    - "Neighborhood"
-    - "RoofStyle"
-    - "MasVnrType"
-    - "BsmtQual"
-    - "BsmtExposure"
-    - "HeatingQC"
-    - "CentralAir"
-    - "KitchenQual"
-    - "FireplaceQu"
-    - "GarageType"
-    - "GarageFinish"
-    - "PavedDrive"
+    - "sex"
+    - "cp"
+    - "restecg"
+    - "exang"
+    - "slope"
+    - "ca"
+    - "thal"
   numerical_features:
-    - "OverallQual"
-    - "MSSubClass"
-    - "OverallCond"
-    - "GrLivArea"
-    - "GarageCars"
-    - "1stFlrSF"
-    - "Fireplaces"
-    - "BsmtFullBath"
-    - "YrSold"
-    - "YearRemodAdd"
-    - "LotFrontage"
+    - "age"
+    - "trestbps"
+    - "chol"
+    - "thalach"
+    - "oldpeak"
   features_to_drop:
-    - "YrSold"
-  target_col: "SalePrice"
-
+    - "fbs"
+  target_col:
+    - "target"
