+++ heart_disease/entities/feature_params.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass, field
+from typing import List, Optional
+
+
+@dataclass()
+class FeatureParams:
+    categorical_features: List[str]
+    numerical_features: List[str]
+    features_to_drop: List[str]
+    target_col: List[str]
+    use_log_trick: bool = field(default=True)
