+++ heart_disease/entities/split_params.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass, field
+
+
+@dataclass()
+class SplittingParams:
+    val_size: float = field(default=0.25)
+    random_state: int = field(default=42)
