+++ heart_disease/entities/train_params.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass, field
+
+
+@dataclass()
+class TrainingParams:
+    model_type: str = field(default="LogisticRegression")
+    random_state: int = field(default=42)
