+++ heart_disease/features/__init__.py
@@ -1,0 +1,3 @@
+from .build_features import make_features
+
+__all__ = ["make_features"]
