+++ heart_disease/models/__init__.py
@@ -1,0 +1,13 @@
+from .model_fit_predict import (
+    train_model,
+    serialize_model,
+    predict_model,
+    evaluate_model,
+)
+
+__all__ = [
+    "train_model",
+    "serialize_model",
+    "evaluate_model",
+    "predict_model",
+]
