+++ heart_disease/models/model_fit_predict.py
@@ -1,0 +1,51 @@
+import pickle
+from typing import Dict, Union
+
+import numpy as np
+import pandas as pd
+from sklearn.ensemble import RandomForestRegressor
+from sklearn.linear_model import LogisticRegression
+from sklearn.metrics import accuracy_score, roc_auc_score
+
+from heart_disease.entities.train_params import TrainingParams
+
+SklearnRegressionModel = Union[RandomForestRegressor, LogisticRegression]
+
+
+def train_model(
+    features: pd.DataFrame, target: pd.Series, train_params: TrainingParams
+) -> SklearnRegressionModel:
+    if train_params.model_type == "RandomForestRegressor":
+        model = RandomForestRegressor(
+            n_estimators=100, random_state=train_params.random_state
+        )
+    elif train_params.model_type == "LogisticRegression":
+        model = LogisticRegression(
+            solver='liblinear', random_state=train_params.random_state
+        )
+    else:
+        raise NotImplementedError()
+    model.fit(features, target)
+    return model
+
+
+def predict_model(
+    model: SklearnRegressionModel, features: pd.DataFrame
+) -> np.ndarray:
+    predicts = model.predict(features)
+    return predicts
+
+
+def evaluate_model(
+    predicts: np.ndarray, target: pd.Series
+) -> Dict[str, float]:
+    return {
+        "accuray": accuracy_score(target, predicts),
+        "roc_auc": roc_auc_score(target, predicts),
+    }
+
+
+def serialize_model(model: SklearnRegressionModel, output: str) -> str:
+    with open(output, "wb") as f:
+        pickle.dump(model, f)
+    return output
