--- ml_example/data/__init__.py
@@ -1,3 +1,0 @@
-from .make_dataset import split_train_val_data, read_data
-
-__all__ = ["split_train_val_data", "read_data"]
