--- ml_example/enities/__init__.py
@@ -1,17 +1,0 @@
-from .feature_params import FeatureParams
-from .split_params import SplittingParams
-from .train_params import TrainingParams
-from .train_pipeline_params import (
-    read_training_pipeline_params,
-    TrainingPipelineParamsSchema,
-    TrainingPipelineParams,
-)
-
-__all__ = [
-    "FeatureParams",
-    "SplittingParams",
-    "TrainingPipelineParams",
-    "TrainingPipelineParamsSchema",
-    "TrainingParams",
-    "read_training_pipeline_params",
-]
