--- ml_example/enities/feature_params.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass, field
-from typing import List, Optional
-
-
-@dataclass()
-class FeatureParams:
-    categorical_features: List[str]
-    numerical_features: List[str]
-    features_to_drop: List[str]
-    target_col: Optional[str]
-    use_log_trick: bool = field(default=True)
