--- ml_example/enities/split_params.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass, field
-
-
-@dataclass()
-class SplittingParams:
-    val_size: float = field(default=0.2)
-    random_state: int = field(default=42)
