--- ml_example/enities/train_params.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass, field
-
-
-@dataclass()
-class TrainingParams:
-    model_type: str = field(default="RandomForestRegressor")
-    random_state: int = field(default=255)
