--- ml_example/features/__init__.py
@@ -1,3 +1,0 @@
-from .build_features import make_features
-
-__all__ = ["make_features"]
