--- ml_example/features/build_features.py
@@ -1,65 +1,0 @@
-import numpy as np
-import pandas as pd
-from sklearn.compose import ColumnTransformer
-from sklearn.impute import SimpleImputer
-from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import OneHotEncoder
-
-from ml_example.enities.feature_params import FeatureParams
-
-
-def process_categorical_features(categorical_df: pd.DataFrame) -> pd.DataFrame:
-
-    categorical_pipeline = build_categorical_pipeline()
-    return pd.DataFrame(categorical_pipeline.fit_transform(categorical_df).toarray())
-
-
-def build_categorical_pipeline() -> Pipeline:
-    categorical_pipeline = Pipeline(
-        [
-            ("impute", SimpleImputer(missing_values=np.nan, strategy="most_frequent")),
-            ("ohe", OneHotEncoder()),
-        ]
-    )
-    return categorical_pipeline
-
-
-def process_numerical_features(numerical_df: pd.DataFrame) -> pd.DataFrame:
-    num_pipeline = build_numerical_pipeline()
-    return pd.DataFrame(num_pipeline.fit_transform(numerical_df))
-
-
-def build_numerical_pipeline() -> Pipeline:
-    num_pipeline = Pipeline(
-        [("impute", SimpleImputer(missing_values=np.nan, strategy="mean")),]
-    )
-    return num_pipeline
-
-
-def make_features(transformer: ColumnTransformer, df: pd.DataFrame) -> pd.DataFrame:
-    return pd.DataFrame(transformer.transform(df).toarray())
-
-
-def build_transformer(params: FeatureParams) -> ColumnTransformer:
-    transformer = ColumnTransformer(
-        [
-            (
-                "categorical_pipeline",
-                build_categorical_pipeline(),
-                params.categorical_features,
-            ),
-            (
-                "numerical_pipeline",
-                build_numerical_pipeline(),
-                params.numerical_features,
-            ),
-        ]
-    )
-    return transformer
-
-
-def extract_target(df: pd.DataFrame, params: FeatureParams) -> pd.Series:
-    target = df[params.target_col]
-    if params.use_log_trick:
-        target = pd.Series(np.log(target.to_numpy()))
-    return target
