+++ requirements.txt
@@ -1,0 +1,7 @@
+setuptools~=52.0.0
+pandas~=1.2.4
+scikit-learn~=0.23.2
+numpy~=1.20.1
+pytest~=6.2.3
+py~=1.10.0
+pyyaml~=5.4.1
