+++ setup.py
@@ -1,19 +1,19 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="ml_example",
+    name="heart_disease",
     packages=find_packages(),
     version="0.1.0",
     description="Example of ml project",
-    author="Your name (or your organization/company/team)",
+    author="Khubbatulin Mark email: markhubbatulin@gmail.com",
     install_requires=[
-        "click==7.1.2",
-        "python-dotenv>=0.5.1",
-        "scikit-learn==0.24.1",
-        "dataclasses==0.8",
-        "pyyaml==3.11",
-        "marshmallow-dataclass==8.3.0",
-        "pandas==1.1.5",
+        "setuptools~=52.0.0",
+        "pandas~=1.2.4",
+        "scikit-learn~=0.23.2",
+        "numpy~=1.20.1",
+        "pytest~=6.2.3",
+        "py~=1.10.0",
+        "pyyaml~=5.4.1"
     ],
     license="MIT",
 )
